requirements.txt
# Database (SQLite is stdlib, LanceDB for vector search)
lancedb>=0.6.0
pyarrow>=14.0

# Machine Learning
sentence-transformers==5.1.0
numpy>=1.21.0

# Google APIs and OAuth
google-api-python-client==2.138.0
google-auth>=2.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# OAuth and Security
aiohttp>=3.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0

# Web and API
requests==2.32.5
flask>=2.0.0

# Utilities
python-dateutil==2.8.2
tqdm>=4.0
